CLEANUP - AutocorrectLogic moved into sole class that includes it
