+++ lib/rubocop.rb
@@ -65,7 +65,6 @@ require_relative 'rubocop/cop/util'
 require_relative 'rubocop/cop/offense'
 require_relative 'rubocop/cop/message_annotator'
 require_relative 'rubocop/cop/ignored_node'
-require_relative 'rubocop/cop/autocorrect_logic'
 require_relative 'rubocop/cop/badge'
 require_relative 'rubocop/cop/registry'
 require_relative 'rubocop/cop/cop'
